refactor for modularity
